Create README.md
● Extracted 2000 employee’s data within an organization and providing insights with business KPI of Total Male & Female, Sunburst visual of region wise employees. ● Top 10 Designation with turnover%, Monthly distribution of Joinee to Leavers count on Stacked Column chart and Total Staff trend and forecast on Line Visual.

Objective:

Help an organization to improve employee performance and improve employee retention (reduce attrition) Monitoring and managing various aspects of employee data to ensure the organization maintains a healthy workforce. Make an organization to assess workforce size and plan for future growth and track the number of employees who have left the organization, hindering the ability to gauge employee satisfaction and engagement, Assessing the current workforce's productivity and capacity, make an organization's ability to attract and retain employees.